Reduce amount of log traces being pushed to Azure Monitor.
