+++ python/device/device.py
@@ -28,13 +28,6 @@ from thief_constants import (
     DeviceSettings as Settings,
 )
 
-logging.basicConfig(level=logging.WARNING)
-logging.getLogger("paho").setLevel(level=logging.DEBUG)
-logging.getLogger("thief").setLevel(level=logging.INFO)
-logging.getLogger("azure.iot").setLevel(level=logging.INFO)
-
-logger = logging.getLogger("thief.{}".format(__name__))
-
 # TODO: exit service when device stops responding
 # TODO: add code to receive rest of pingacks at end.  wait for delta since last to be > 20 seconds.
 # TODO: add mid to debug logs as custom property, maybe service_ack id
@@ -48,6 +41,14 @@ requested_service_pool = os.environ["THIEF_REQUESTED_SERVICE_POOL"]
 
 run_id = str(uuid.uuid4())
 
+# set default logging which will only go to the console
+
+logging.basicConfig(level=logging.WARNING)
+logging.getLogger("paho").setLevel(level=logging.DEBUG)
+logging.getLogger("thief").setLevel(level=logging.INFO)
+logging.getLogger("azure.iot").setLevel(level=logging.INFO)
+logger = logging.getLogger("thief.{}".format(__name__))
+
 # configure our traces and events to go to Azure Monitor
 azure_monitor.add_logging_properties(
     client_type="device",
@@ -57,9 +58,8 @@ azure_monitor.add_logging_properties(
     pool_id=requested_service_pool,
 )
 event_logger = azure_monitor.get_event_logger()
+azure_monitor.log_all_warnings_and_exceptions_to_azure_monitor()
 azure_monitor.log_to_azure_monitor("thief")
-azure_monitor.log_to_azure_monitor("azure")
-azure_monitor.log_to_azure_monitor("paho")
 
 
 class ServiceAckWaitInfo(object):
@@ -505,6 +505,7 @@ class DeviceApp(app_base.AppBase):
                     )
 
                 else:
+                    azure_monitor.add_logging_properties(service_instance=received_service_instance)
                     # It looks like a service app has decided to pair with us.  Set reported
                     # properties to "select" this service instance as our partner.
                     logger.info(
